Title: 소괄호 매칭 수 최대화하기, Time: 37ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 240302/소괄호 매칭 수 최대화하기/maximize-the-number-of-parenthesis-matches.cpp
@@ -27,7 +27,7 @@ int main() {
         string str = arr;
 
         int open = 0, close = 0;
-        for (int j = 0; j < str.length(); j++) {
+        for (int j = 0; j < (int) str.length(); j++) {
             if (str[j] == '(') {
                 open++;
             }
